+++ go-pinot-api.go
@@ -464,40 +464,3 @@ func (c *PinotAPIClient) DeleteSchema(schemaName string) (*model.UserActionRespo
 func fullUrl(url *url.URL, path string) string {
 	return fmt.Sprintf("http://%s:%s%s", url.Hostname(), url.Port(), path)
 }
-
-func validateOpts(opts ...Opt) (*cfg, *url.URL, error) {
-
-	// with default auth writer that does nothing
-	parsedCfg := defaultCfg()
-	for _, opt := range opts {
-		opt.apply(parsedCfg)
-	}
-
-	// validate controller url
-	pinotControllerUrl, err := url.Parse(parsedCfg.controllerUrl)
-	if err != nil {
-		return nil, nil, fmt.Errorf("controller url is invalid: %w", err)
-	}
-
-	// if auth token passed, handle authenticated requests
-	if parsedCfg.authToken != "" {
-		parsedCfg.httpAuthWriter = func(req *http.Request) {
-			req.Header.Set("Authorization", fmt.Sprintf("Basic %s", parsedCfg.authToken))
-		}
-	}
-
-	return parsedCfg, pinotControllerUrl, nil
-
-}
-
-func defaultCfg() *cfg {
-	return &cfg{
-		httpAuthWriter: defaultAuthWriter(),
-	}
-}
-
-func defaultAuthWriter() func(*http.Request) {
-	return func(req *http.Request) {
-		// do nothing
-	}
-}
